+++ test/integration/test_integration_rioxarray.py
@@ -4,6 +4,7 @@ import numpy
 import pytest
 import rasterio
 import xarray
+from affine import Affine
 from numpy.testing import assert_almost_equal, assert_array_equal
 from rasterio.crs import CRS
 
@@ -261,10 +262,9 @@ def test_clip_geojson():
         # add transform for test
         comp_subset.attrs["transform"] = tuple(comp_subset.rio.transform(recalc=True))
         # add grid mapping for test
-        comp_subset.attrs["grid_mapping"] = "spatial_ref"
-        comp_subset.coords["spatial_ref"] = 0
+        comp_subset.rio.write_crs(subset.rio.crs, inplace=True)
         # make sure nodata exists for test
-        comp_subset.attrs["_FillValue"] = comp_subset.attrs["nodatavals"][0]
+        comp_subset.attrs["_FillValue"] = comp_subset.rio.nodata
 
         geometries = [
             {
@@ -282,7 +282,7 @@ def test_clip_geojson():
         ]
 
         # test data array
-        clipped = xdi.rio.clip(geometries, subset.rio.crs)
+        clipped = xdi.rio.clip(geometries, comp_subset.rio.crs)
         _assert_xarrays_equal(clipped, comp_subset)
 
         # test dataset
@@ -291,6 +291,39 @@ def test_clip_geojson():
         )
         comp_subset_ds = comp_subset.to_dataset(name="test_data")
         _assert_xarrays_equal(clipped_ds, comp_subset_ds)
+
+
+def test_clip_geojson__no_drop():
+    with xarray.open_rasterio(
+        os.path.join(TEST_COMPARE_DATA_DIR, "small_dem_3m_merged.tif")
+    ) as xdi:
+        geometries = [
+            {
+                "type": "Polygon",
+                "coordinates": [
+                    [
+                        [-93.880889448126, 41.68465068553298],
+                        [-93.89966980835203, 41.68465068553298],
+                        [-93.89966980835203, 41.689430423525266],
+                        [-93.880889448126, 41.689430423525266],
+                        [-93.880889448126, 41.68465068553298],
+                    ]
+                ],
+            }
+        ]
+        # test data array
+        clipped = xdi.rio.clip(geometries, "epsg:4326", drop=False)
+        assert clipped.rio.crs == xdi.rio.crs
+        assert clipped.shape == xdi.shape
+        assert clipped.sum().item() == 2150801411
+
+        # test dataset
+        clipped_ds = xdi.to_dataset(name="test_data").rio.clip(
+            geometries, "epsg:4326", drop=False
+        )
+        assert clipped_ds.rio.crs == xdi.rio.crs
+        assert clipped_ds.test_data.shape == xdi.shape
+        assert clipped_ds.test_data.sum().item() == 2150801411
 
 
 def test_transform_bounds():
@@ -727,6 +760,40 @@ def test_to_raster_3d(tmpdir):
         assert_array_equal(rds.transform, xds.rio.transform())
         assert_array_equal(rds.nodata, xds.rio.nodata)
         assert_array_equal(rds.read(), xds.values)
+
+
+def test_to_raster__preserve_profile__none_nodata(tmpdir):
+    tmp_raster = tmpdir.join("output_profile.tif")
+    input_raster = tmpdir.join("input_profile.tif")
+
+    transform = Affine.from_gdal(0, 512, 0, 0, 0, 512)
+    with rasterio.open(
+        str(input_raster),
+        "w",
+        driver="GTiff",
+        height=512,
+        width=512,
+        count=1,
+        crs="+init=epsg:4326",
+        transform=transform,
+        dtype=rasterio.float32,
+        tiled=True,
+        tilexsize=256,
+        tileysize=256,
+    ) as rds:
+        rds.write(numpy.empty((1, 512, 512), dtype=numpy.float32))
+
+    with xarray.open_rasterio(str(input_raster)) as mda:
+        mda.rio.to_raster(str(tmp_raster))
+
+    with rasterio.open(str(tmp_raster)) as rds, rasterio.open(str(input_raster)) as rdc:
+        assert rds.count == rdc.count
+        assert rds.crs == rdc.crs
+        assert_array_equal(rds.transform, rdc.transform)
+        assert_array_equal(rds.nodata, rdc.nodata)
+        assert_array_equal(rds.read(), rdc.read())
+        assert rds.profile == rdc.profile
+        assert rds.nodata is None
 
 
 def test_missing_dimensions():